requirements.txt
jax==0.2.10
jaxlib==0.1.62
flax==0.3.0
numpyro==0.6.0
ml-collections==0.1.0
absl-py==1.0.0
tensorflow
tensorflow_datasets
tensorboard